fix: fix alignment of text in bale legends
All the text is a `<tspan>` inside a `<text>`, both with the same `x` and
`y` attributes for positioning.  I'm not sure which position matters,
but only the position for the `<text>` was being re-computed based on
bale size — apparently that used to work (with older Inkscapes), but
no longer.  Now, just set both positions.